Refactor to allow multiple Bolt protocol versions Added factory method for protocol selection
