+++ Neo4j.Driver/Neo4j.Driver/Internal/Connector/IBoltProtocol.cs
@@ -1,34 +1,33 @@
-// Copyright (c) 2002-2018 "Neo Technology,"
+﻿// Copyright (c) 2002-2018 "Neo Technology,"
 // Network Engine for Objects in Lund AB [http://neotechnology.com]
-// 
+//
 // This file is part of Neo4j.
-// 
+//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-using System;
-using System.IO;
-using System.Threading.Tasks;
+
+using Neo4j.Driver.Internal.IO;
 
 namespace Neo4j.Driver.Internal.Connector
 {
-    internal interface ITcpSocketClient : IDisposable
+    internal interface IBoltProtocol
     {
-        Stream ReadStream { get;  }
-        Stream WriteStream { get;  }
-
-        void Connect(Uri uri);
-        Task ConnectAsync(Uri uri);
-
-        void Disconnect();
-        Task DisconnectAsync();
+        IBoltReader Reader { get; }
+        IBoltWriter Writer { get; }
+        /// <summary>
+        /// Reconfig protocol if necessary.
+        /// </summary>
+        /// <param name="serverVersion"></param>
+        /// <returns>true if reconfigured, otherwise false.</returns>
+        bool ReconfigIfNecessary(string serverVersion);
     }
 }
