+++ Neo4j.Driver/Neo4j.Driver/Internal/Connector/SocketClient.cs
@@ -27,17 +27,9 @@ namespace Neo4j.Driver.Internal.Connector
 {
     internal class SocketClient :  ISocketClient
     {
-        private static class ProtocolVersion
-        {
-            public const int NoVersion = 0;
-            public const int Version1 = 1;
-            public const int Http = 1213486160;
-        }
-
         private readonly ITcpSocketClient _tcpSocketClient;
         private readonly Uri _uri;
-        private IBoltReader _reader;
-        private IBoltWriter _writer;
+        private IBoltProtocol _boltProtocol;
         private readonly BufferSettings _bufferSettings;
 
         private int _closedMarker = -1;
@@ -64,10 +56,9 @@ namespace Neo4j.Driver.Internal.Connector
             _logger?.Debug($"~~ [CONNECT] {_uri}");
 
             var version = DoHandshake();
-            
-            ConfigureVersion(version);
+            _boltProtocol = BoltProtocolFactory.Create(version, _tcpSocketClient, _bufferSettings, _logger);
         }
-        
+
         public Task StartAsync()
         {
             TaskCompletionSource<object> tcs = new TaskCompletionSource<object>();
@@ -101,8 +92,7 @@ namespace Neo4j.Driver.Internal.Connector
                     {
                         try
                         {
-                            ConfigureVersion(version);
-
+                            _boltProtocol = BoltProtocolFactory.Create(version, _tcpSocketClient, _bufferSettings, _logger);
                             tcs.SetResult(null);
                         }
                         catch (AggregateException exc)
@@ -119,22 +109,16 @@ namespace Neo4j.Driver.Internal.Connector
             return tcs.Task;
         }
 
-        private void SetupPackStreamFormatWriterAndReader(bool supportBytes = true)
-        {
-            _writer = new BoltWriter(_tcpSocketClient.WriteStream, _bufferSettings.DefaultWriteBufferSize, _bufferSettings.MaxWriteBufferSize, _logger, supportBytes); 
-            _reader = new BoltReader(_tcpSocketClient.ReadStream, _bufferSettings.DefaultReadBufferSize, _bufferSettings.MaxReadBufferSize, _logger, supportBytes);
-        }
-        
         public void Send(IEnumerable<IRequestMessage> messages)
         {
             try
             {
                 foreach (var message in messages)
                 {
-                    _writer.Write(message);
+                    _boltProtocol.Writer.Write(message);
                     _logger?.Debug("C: ", message);
                 }
-                _writer.Flush();
+                _boltProtocol.Writer.Flush();
             }
             catch (Exception ex)
             {
@@ -152,11 +136,11 @@ namespace Neo4j.Driver.Internal.Connector
             {
                 foreach (var message in messages)
                 {
-                    _writer.Write(message);
+                    _boltProtocol.Writer.Write(message);
                     _logger?.Debug("C: ", message);
                 }
 
-                _writer.FlushAsync().ContinueWith(t =>
+                _boltProtocol.Writer.FlushAsync().ContinueWith(t =>
                 {
                     if (t.IsFaulted && t.Exception != null)
                     {
@@ -204,7 +188,7 @@ namespace Neo4j.Driver.Internal.Connector
         {
             try
             {
-                _reader.Read(responseHandler);
+                _boltProtocol.Reader.Read(responseHandler);
             }
             catch (Exception ex)
             {
@@ -224,7 +208,7 @@ namespace Neo4j.Driver.Internal.Connector
         {
             TaskCompletionSource<object> tcs = new TaskCompletionSource<object>();
 
-            _reader.ReadAsync(responseHandler).ContinueWith(t =>
+            _boltProtocol.Reader.ReadAsync(responseHandler).ContinueWith(t =>
             {
                 if (t.IsFaulted)
                 {
@@ -257,90 +241,16 @@ namespace Neo4j.Driver.Internal.Connector
             return tcs.Task;
         }
 
+        public void UpdateBoltProtocol(string serverVersion)
+        {
+            _boltProtocol.ReconfigIfNecessary(serverVersion);
+        }
+
         private void SetOpened()
         {
             Interlocked.CompareExchange(ref _closedMarker, 0, -1);
         }
 
-        private int DoHandshake()
-        {
-            int[] supportedVersion = {1, 0, 0, 0};
-            
-            var data = PackVersions(supportedVersion);
-            _tcpSocketClient.WriteStream.Write(data, 0, data.Length);
-            _tcpSocketClient.WriteStream.Flush();
-            _logger?.Debug("C: [HANDSHAKE] [0x6060B017, 1, 0, 0, 0]");
-
-            data = new byte[4];
-            _tcpSocketClient.ReadStream.Read(data, 0, data.Length);
-
-            var agreedVersion = GetAgreedVersion(data);
-            _logger?.Debug($"S: [HANDSHAKE] {agreedVersion}");
-            return agreedVersion;
-        }
-
-        private async Task<int> DoHandshakeAsync()
-        {
-            int[] supportedVersion = {1, 0, 0, 0};
-            
-            var data = PackVersions(supportedVersion);
-            await _tcpSocketClient.WriteStream.WriteAsync(data, 0, data.Length).ConfigureAwait(false);
-            await _tcpSocketClient.WriteStream.FlushAsync().ConfigureAwait(false);
-            _logger?.Debug("C: [HANDSHAKE] [0x6060B017, 1, 0, 0, 0]");
-
-            data = new byte[4];
-            await _tcpSocketClient.ReadStream.ReadAsync(data, 0, data.Length).ConfigureAwait(false);
-
-            var agreedVersion = GetAgreedVersion(data);
-            _logger?.Debug($"S: [HANDSHAKE] {agreedVersion}");
-            return agreedVersion;
-        }
-
-        private void ConfigureVersion(int version)
-        {
-            switch (version)
-            {
-                case ProtocolVersion.Version1:
-                    SetupPackStreamFormatWriterAndReader();
-                    break;
-                case ProtocolVersion.NoVersion:
-                    throw new NotSupportedException(
-                        "The Neo4j server does not support any of the protocol versions supported by this client. " +
-                        "Ensure that you are using driver and server versions that are compatible with one another.");
-                case ProtocolVersion.Http:
-                    throw new NotSupportedException(
-                        "Server responded HTTP. Make sure you are not trying to connect to the http endpoint " +
-                        $"(HTTP defaults to port 7474 whereas BOLT defaults to port {GraphDatabase.DefaultBoltPort})");
-                default:
-                    throw new NotSupportedException(
-                        "Protocol error, server suggested unexpected protocol version: " + version);
-            }
-        }
-
-        private static byte[] PackVersions(IEnumerable<int> versions)
-        {
-            //This is a 'magic' handshake identifier to indicate we're using 'BOLT' ('GOGOBOLT')
-            var aLittleBitOfMagic = PackStreamBitConverter.GetBytes(0x6060B017);
-
-            var bytes = new List<byte>(aLittleBitOfMagic);
-            foreach (var version in versions)
-            {
-                bytes.AddRange(PackStreamBitConverter.GetBytes(version));
-            }
-            return bytes.ToArray();
-        }
-
-
-        private static int GetAgreedVersion(byte[] data)
-        {
-            return PackStreamBitConverter.ToInt32(data);
-        }
-
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
 
         public void Stop()
         {
@@ -360,6 +270,12 @@ namespace Neo4j.Driver.Internal.Connector
             return TaskExtensions.GetCompletedTask();
         }
 
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
         protected virtual void Dispose(bool disposing)
         {
             if (IsClosed)
@@ -371,14 +287,34 @@ namespace Neo4j.Driver.Internal.Connector
             }
         }
 
-        public void UpdatePackStream(string serverVersion)
+        private int DoHandshake()
         {
-            var version = ServerVersion.Version(serverVersion);
-            if ( version >= ServerVersion.V3_2_0 )
-            {
-                return;
-            }
-            SetupPackStreamFormatWriterAndReader(false);
+            var data = BoltProtocolFactory.PackSupportedVersions();
+            _tcpSocketClient.WriteStream.Write(data, 0, data.Length);
+            _tcpSocketClient.WriteStream.Flush();
+            _logger?.Debug("C: [HANDSHAKE] ", data);
+
+            data = new byte[4];
+            _tcpSocketClient.ReadStream.Read(data, 0, data.Length);
+
+            var agreedVersion = BoltProtocolFactory.UnpackAgreedVersion(data);
+            _logger?.Debug($"S: [HANDSHAKE] {agreedVersion}");
+            return agreedVersion;
+        }
+
+        private async Task<int> DoHandshakeAsync()
+        {
+            var data = BoltProtocolFactory.PackSupportedVersions();
+            await _tcpSocketClient.WriteStream.WriteAsync(data, 0, data.Length).ConfigureAwait(false);
+            await _tcpSocketClient.WriteStream.FlushAsync().ConfigureAwait(false);
+            _logger?.Debug("C: [HANDSHAKE] ", data);
+
+            data = new byte[4];
+            await _tcpSocketClient.ReadStream.ReadAsync(data, 0, data.Length).ConfigureAwait(false);
+
+            var agreedVersion = BoltProtocolFactory.UnpackAgreedVersion(data);
+            _logger?.Debug($"S: [HANDSHAKE] {agreedVersion}");
+            return agreedVersion;
         }
     }
 }
